refactor: extract MCP tool handlers to shared module
Move handleToolsList() and handleToolCall() plus all their internal
helpers (validate, handleProxy, handleLocal, handleProxyGeneric,
normalizeList, parseLimit, textResponse, announceMode) from
src/mcp/standalone.ts into a new src/mcp/handler.ts shared module.

The extracted functions accept explicit sdk, kv, and config parameters
so they can be reused by both the existing stdio transport and the
upcoming Stream HTTP transport without code duplication.

standalone.ts retains:
- InMemoryKV instantiation
- createStdioTransport and the stdio message dispatch loop
- SIGINT/SIGTERM handlers
- Backward-compatible wrapper exports (old signatures) for tests

All 1410 non-integration tests pass; no behavioral changes.
